+++ memo/ChangeLog.md
@@ -45,6 +45,7 @@
 - prompt: support `bleopt prompt_{xterm_title,screen_title,status_line}` `#D1378` 5c3f6fe
   - prompt: check `TERM` for prompt window titles when `_ble_term_TERM` is unavailable `#D1388` 3c88869
 - syntax: support options `bleopt highlight_{syntax,filename,vartype}` to turn off highlighting (requested by pjmp) `#D1379` 0116f8b
+- complete: support `shopt progcomp_alias` `#D1397` 0000000
 
 ## Changes